Fixed issue on favEvent page
