+++ src/app/favlist/styles.ts
@@ -6,5 +6,6 @@ export const useStyles = createStyles(() => ({
     alignItems: "center",
     flexDirection: "column",
     paddingTop: "5%",
+    textAlign: "center",
   },
 }));
